backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4